2_Onnx2Nef/requirements.txt
onnx
pillow